Add a high-level hit test report for bevy_iced.
The idea is to simulate `wants_pointer_input` in egui. Implementation of this is primitive (ha ha) and hacky, based on the final primitive list supplied to the renderer, and is unlikely to last forever, but is surprisingly effective as a first-order approximation. For example, it will ignore containers (rows, columns, etc.) without borders or backgrounds, because they don't generate quads, and only look at the content inside them.

This is "good enough" to be used as a run condition for now, as long as care is taken to e.g. not initiate dragging unless a real click was received.